If exposing an NSDictionary inside v8, ensure changes are mirrored in ObjC-land.
This allows for dot notation in JS such as `dictionary.prop1.prop2 = true`.
Without something writing back to the mutable dictionary in ObjC-land,
the named getter will just keep returning a new v8 object instance of the original 
ObjC dictionary.
Cannot determine if the dictionary is mutable (class cluster) so this applies to all.
The same needs to be done for v8 arrays.